fix: Workaround Intel GPU issues in multi-draw renderer
Should have minimal performance impact, if any. It seems Intel's
drivers on Windows can't handle accessing indirect commands from
a buffer and needs the command array passed as a pointer into
client memory.

Of course, AMD requires the opposite, because the state of OpenGL
drivers on Windows is nothing but sadness.